Software/back-end/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.62.0
h11==0.11.0
idna==2.10
iniconfig==1.1.1
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
urllib3==1.26.2
uvicorn==0.13.1
validate-docbr==1.8.2
watchgod==0.6
<<<<<<< HEAD
websockets==8.1
=======
websockets==8.1
>>>>>>> ffff7d25f1448e797d7f5cec993cbf0bcf51e352
